+++ src/main.js
@@ -1,7 +1,7 @@
 import { createApp } from 'vue';
 import App from './App.vue';
 import router from './router';
-import store from './store';
+import store from './store/store';
 
 const app = createApp(App);
 app.use(router);
