+++ src/store/getters.js
@@ -1,0 +1,10 @@
+export default {
+    gender: state => state.gender,
+    year: state => state.year,
+    month: state => state.month,
+    day: state => state.day,
+    question1: state => state.question1,
+    question2: state => state.question2,
+    question3: state => state.question3,
+    consultationContent: state => state.consultationContent
+}
