+++ src/store/mutations.js
@@ -1,0 +1,26 @@
+export default {
+    updateGender(state, newGender) {
+        state.gender = newGender;
+    },
+    updateYear(state, newYear) {
+        state.year = newYear;
+    },
+    updateMonth(state, newMonth) {
+        state.month = newMonth;
+    },
+    updateDay(state, newDay) {
+        state.day = newDay;
+    },
+    updateQuestion1(state, newQuestion1) {
+        state.question1 = newQuestion1;
+    },
+    updateQuestion2(state, newQuestion2) {
+        state.question2 = newQuestion2;
+    },
+    updateQuestion3(state, newQuestion3) {
+        state.question3 = newQuestion3;
+    },
+    updateConsultationContent(state, newConsultationContent) {
+        state.consultationContent = newConsultationContent;
+    }
+}
